pyproject.toml
[tool.poetry]
name = "pharmaai"
version = "1.0.0"
description = "PharmaAI Assistant – Pharmaceutical Intelligence Platform"
authors = ["PharmaAI Team"]
readme = "README.md"
python = "^3.11"

[tool.poetry.dependencies]
python = "^3.11"

# ── Deep Learning ────────────────────────────────────────────────────────────
torch = {version = "^2.2.0", extras = ["cuda"]}
transformers = "^4.40.0"
accelerate = "^0.30.0"
peft = "^0.10.0"           # LoRA / QLoRA
bitsandbytes = "^0.43.0"   # 4-bit / 8-bit quantisation
datasets = "^2.19.0"
sentencepiece = "^0.2.0"
einops = "^0.8.0"

# ── Vector Index ─────────────────────────────────────────────────────────────
faiss-gpu = "^1.7.4"
annoy = "^1.17.3"

# ── API Framework ─────────────────────────────────────────────────────────────
fastapi = "^0.111.0"
uvicorn = {version = "^0.29.0", extras = ["standard"]}
pydantic = "^2.7.0"
pydantic-settings = "^2.2.0"
httpx = "^0.27.0"
aiohttp = "^3.9.0"

# ── Databases ─────────────────────────────────────────────────────────────────
asyncpg = "^0.29.0"
sqlalchemy = {version = "^2.0.0", extras = ["asyncio"]}
alembic = "^1.13.0"
redis = {version = "^5.0.0", extras = ["asyncio"]}
motor = "^3.4.0"           # async MongoDB (optional)

# ── Message Queue ─────────────────────────────────────────────────────────────
aio-pika = "^9.4.0"
celery = {version = "^5.4.0", extras = ["redis"]}

# ── NLP / Text ────────────────────────────────────────────────────────────────
nltk = "^3.8.1"
spacy = "^3.7.0"
rank-bm25 = "^0.2.2"
sumy = "^0.11.0"

# ── Rasa ──────────────────────────────────────────────────────────────────────
rasa = "^3.6.0"

# ── Notifications ─────────────────────────────────────────────────────────────
twilio = "^9.0.0"
slack-sdk = "^3.27.0"
sendgrid = "^6.11.0"
firebase-admin = "^6.5.0"

# ── Monitoring ────────────────────────────────────────────────────────────────
prometheus-client = "^0.20.0"
prometheus-fastapi-instrumentator = "^7.0.0"
opentelemetry-sdk = "^1.24.0"
opentelemetry-instrumentation-fastapi = "^0.45b0"

# ── Utilities ─────────────────────────────────────────────────────────────────
pyyaml = "^6.0.1"
python-dotenv = "^1.0.1"
python-json-logger = "^2.0.7"
tenacity = "^8.3.0"
cachetools = "^5.3.3"
structlog = "^24.1.0"
rich = "^13.7.1"
typer = "^0.12.0"
tqdm = "^4.66.0"
numpy = "^1.26.0"
pandas = "^2.2.0"
scipy = "^1.13.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"
httpx = "^0.27.0"         # for FastAPI test client
black = "^24.4.0"
ruff = "^0.4.0"
mypy = "^1.10.0"
pre-commit = "^3.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=pharmaai --cov-report=term-missing"